Move logical Editor into /api module folder
