+++ code/ui.datagrid/src/api/index.ts
@@ -1,2 +1,3 @@
 export * from './Grid';
 export * from './Cell';
+export * from './Editor';
