+++ code/ui.datagrid/src/types.ts
@@ -1,7 +1,7 @@
 export { KeypressObservable } from '@platform/react';
 
+export * from './api/Editor/types';
 export * from './components/DataGrid/types';
-export * from './components/Editor/types';
 export * from './components/render/types';
 export * from './components/factory/types';
